Intercept and validate update requests
