+++ test/e2e/basic_test.go
@@ -13,12 +13,15 @@ package e2e
 import (
 	"testing"
 
+	. "github.com/onsi/gomega"
 	framework "github.com/operator-framework/operator-sdk/pkg/test"
 	pravega_e2eutil "github.com/pravega/pravega-operator/pkg/test/e2e/e2eutil"
 )
 
 // Test create and recreate a Pravega cluster with the same name
 func testCreateRecreateCluster(t *testing.T) {
+	g := NewGomegaWithT(t)
+
 	doCleanup := true
 	ctx := framework.NewTestCtx(t)
 	defer func() {
@@ -28,67 +31,45 @@ func testCreateRecreateCluster(t *testing.T) {
 	}()
 
 	namespace, err := ctx.GetNamespace()
-	if err != nil {
-		t.Fatal(err)
-	}
+	g.Expect(err).NotTo(HaveOccurred())
 	f := framework.Global
 
 	defaultCluster := pravega_e2eutil.NewDefaultCluster(namespace)
 
 	pravega, err := pravega_e2eutil.CreateCluster(t, f, ctx, defaultCluster)
-	if err != nil {
-		t.Fatal(err)
-	}
+	g.Expect(err).NotTo(HaveOccurred())
 
 	// A default Pravega cluster should have 5 pods: 3 bookies, 1 controller, 1 segment store
 	podSize := 5
 	err = pravega_e2eutil.WaitForClusterToBecomeReady(t, f, ctx, pravega, podSize)
-	if err != nil {
-		t.Fatal(err)
-	}
+	g.Expect(err).NotTo(HaveOccurred())
 
 	err = pravega_e2eutil.DeleteCluster(t, f, ctx, pravega)
-	if err != nil {
-		t.Fatal(err)
-	}
+	g.Expect(err).NotTo(HaveOccurred())
 
 	err = pravega_e2eutil.WaitForClusterToTerminate(t, f, ctx, pravega)
-	if err != nil {
-		t.Fatal(err)
-	}
+	g.Expect(err).NotTo(HaveOccurred())
 
 	defaultCluster = pravega_e2eutil.NewDefaultCluster(namespace)
 
 	pravega, err = pravega_e2eutil.CreateCluster(t, f, ctx, defaultCluster)
-	if err != nil {
-		t.Fatal(err)
-	}
+	g.Expect(err).NotTo(HaveOccurred())
 
 	err = pravega_e2eutil.WaitForClusterToBecomeReady(t, f, ctx, pravega, podSize)
-	if err != nil {
-		t.Fatal(err)
-	}
+	g.Expect(err).NotTo(HaveOccurred())
 
 	err = pravega_e2eutil.WriteAndReadData(t, f, ctx, pravega)
-	if err != nil {
-		t.Fatal(err)
-	}
+	g.Expect(err).NotTo(HaveOccurred())
 
 	err = pravega_e2eutil.DeleteCluster(t, f, ctx, pravega)
-	if err != nil {
-		t.Fatal(err)
-	}
+	g.Expect(err).NotTo(HaveOccurred())
 
 	// No need to do cleanup since the cluster CR has already been deleted
 	doCleanup = false
 
 	err = pravega_e2eutil.WaitForClusterToTerminate(t, f, ctx, pravega)
-	if err != nil {
-		t.Fatal(err)
-	}
+	g.Expect(err).NotTo(HaveOccurred())
 
 	err = pravega_e2eutil.RestartTier2(t, f, ctx, namespace)
-	if err != nil {
-		t.Fatal(err)
-	}
+	g.Expect(err).NotTo(HaveOccurred())
 }
